requirements.txt
streamlit 
streamlit-chat
pandasai==2.2.12
numpy==1.26.4
pandas==1.5.3
pyyaml==6.0.1
google-generativeai